nutri-backend/requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers>=4.3.1
drf-spectacular==0.26.5
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
gunicorn>=21.2.0
python-dotenv>=1.0.1
Pillow>=10.2.0
uuid>=1.30
drf-spectacular>=0.27.1
